Refactor package structure for POS refund-related classes
