+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/woopos/root/navigation/WooPosMainFlowGraph.kt
@@ -7,8 +7,8 @@ import com.woocommerce.android.ui.woopos.emailreceipt.emailReceiptScreen
 import com.woocommerce.android.ui.woopos.home.WooPosHomeViewModel
 import com.woocommerce.android.ui.woopos.home.eligibilityScreen
 import com.woocommerce.android.ui.woopos.home.homeScreen
+import com.woocommerce.android.ui.woopos.orders.details.refund.refundReasonScreen
 import com.woocommerce.android.ui.woopos.orders.ordersScreen
-import com.woocommerce.android.ui.woopos.orders.refundReasonScreen
 import com.woocommerce.android.ui.woopos.settings.settingsScreen
 import com.woocommerce.android.ui.woopos.splash.SPLASH_ROUTE
 import com.woocommerce.android.ui.woopos.splash.splashScreen
